+++ chringo/lib/src/records/wordbanks_record.dart
