+++ chringo/lib/src/widgets/bingo_grid.dart
@@ -1,6 +1,8 @@
 import 'dart:math';
+
 import 'package:flutter/material.dart';
-import 'bingo_tile.dart';
+
+import '../widgets/bingo_tile.dart';
 
 class BingoGrid extends StatelessWidget {
   final List<String> words;
